src/requirements.txt
Django==1.9.5
Fabric==1.11.1
pycrypto==2.6.1
paramiko==1.17.0
Markdown==2.6.6
Pillow==3.2.0
Unipath==1.1
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
boto==2.40.0
celery==3.1.23
kombu>=3.0
pytz==2016.4
django-allauth==0.25.2
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
django-appconf==1.0.2
django-celery==3.1.17
django-compressor==2.0
django-cors-headers==1.1.0
django-debug-toolbar==1.4
sqlparse==0.1.19
django-devserver==0.8.0
django-grappelli==2.8.1
django-filebrowser==3.7.1
django-filter==0.13.0
django-phonenumber-field==1.1.0
django-registration-redux==1.4
djangorestframework==3.3.3
PyYAML==3.11
django-rest-swagger==0.3.6
django-reversion==1.10.2
django-ses==0.7.1
django-storages==1.4.1
django-twilio==0.8.0
httplib2==0.9.2
twilio==5.4.0
ecdsa==0.9
gevent==1.1.1
greenlet==0.4.9
gunicorn==19.4.5
meld3==1.0.2
psycopg2==2.6.1
pygeocoder==1.2.5
pylibmc==1.5.1
python-dateutil==2.5.3
simplejson==3.8.2
six==1.7.3
supervisor==3.2.3
uWSGI==2.0.12
unittest2==1.1.0
uwsgitop==0.9
wsgiref==0.1.2
oauthlib>=1.0.0
django-ajax-selects==1.4.3